+++ package.json
@@ -14,13 +14,14 @@
         "prettier": "^2.8.0",
         "prettier-plugin-tailwindcss": "^0.2.0",
         "tailwindcss": "^3.0.17",
-        "vite": "^3.2.4",
+        "vite": "^3.2.7",
         "vue-loader": "^16.8.3"
     },
     "dependencies": {
         "@sentry/vue": "^7.23.0",
+        "@typeform/embed": "^2.12.1",
         "@vitejs/plugin-vue": "^3.2.0",
-        "@vueuse/core": "^7.7.1",
+        "@vueuse/core": "^10.2.0",
         "laravel-vue-i18n": "^1.2.0",
         "pinia": "^2.0.11",
         "vue": "^3.2.29",
