+++ resources/js/views/Interaction.vue
@@ -51,9 +51,9 @@
 </style>
 
 <script setup>
-import { nextTick, onMounted, watch, onUnmounted, ref } from 'vue'
+import { nextTick, onMounted } from 'vue'
+import { watchDebounced } from '@vueuse/core'
 import { useRoute } from 'vue-router'
-import { storeToRefs } from 'pinia'
 import CodePanel from '../components/CodePanel.vue'
 import InteractionItemFavourited from '../components/InteractionItemFavourited.vue'
 import InteractionItemViewed from '../components/InteractionItemViewed.vue'
@@ -65,7 +65,6 @@ import { useItemStore } from '../stores/ItemStore'
 import { useSectionStore } from '../stores/SectionStore'
 import { useStoryStore } from '../stores/StoryStore'
 import { usePlaceStore } from '../stores/PlaceStore'
-import { useSurveyStore } from '../stores/SurveyStore'
 import { useSurvey } from '../composables/Survey'
 
 const interactionStore = useInteractionStore()
@@ -126,21 +125,16 @@ onMounted(async () => {
 
     nextTick(scrollActiveIntoView)
 })
-const surveyTimeoutId = ref()
 
-onUnmounted(() => {
-    clearTimeout(surveyTimeoutId.value)
-})
+const { toggle: toggleSurvey, shouldLaunch: shouldSurveyLaunch } = useSurvey()
 
-const store = useSurveyStore()
-const { toggle: toggleSurvey } = useSurvey()
-const { shouldLaunch } = storeToRefs(store)
-
-watch(shouldLaunch, () => {
-    if (shouldLaunch.value) {
-        surveyTimeoutId.value = setTimeout(toggleSurvey, 1000)
-    }
-}, { immediate: true })
-
-
+watchDebounced(
+    shouldSurveyLaunch,
+    () => {
+        if (shouldSurveyLaunch.value) {
+            toggleSurvey()
+        }
+    },
+    { immediate: true, debounce: 1000 }
+)
 </script>
